% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_datasets.R
\docType{data}
\name{available_datasets}
\alias{available_datasets}
\title{Available Datasets}
\format{
A data frame with 7 rows and 5 variables.
\describe{
\item{dataset}{Dataset name used by the WHO data dictionary.}
\item{description}{Either the data description supplied by the WHO or a user generated description.}
\item{timespan}{The timespan of the data}
\item{default}{Whether the dataset is downloaded by default or not.}
\item{url}{The URL for downloading the data - used by \code{\link[getTBinR]{get_tb_burden}}}
}
}
\usage{
available_datasets
}
\description{
Datasets that \code{getTBinR} supports importing into R, see \href{https://www.who.int/tb/country/data/download/en/}{here}
additional dataset details. This dataset is also used by \code{\link[getTBinR]{get_tb_burden}}.
Use \code{\link[getTBinR]{search_data_dict}} in order to get details of the variables in each dataset.
}
\keyword{datasets}
