% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Generic Get Data Function}
\usage{
get_data(url = NULL, data_trans_fn = NULL, path = "data-raw",
  save = TRUE, save_name = NULL, return = TRUE)
}
\arguments{
\item{url}{Character string, indicating  the url of the data to download.}

\item{data_trans_fn}{Function that takes a data.table as input and returns a single
dataframe of any type. If not specified defaults to transforming the data into a tibble.}

\item{path}{Character string, indicating the folder to save the data into,
defaults to \code{data-raw}.}

\item{save}{Logical, should the data be saved for reuse. Defaults to
\code{TRUE}.}

\item{save_name}{Character string, name to save the data under. Defaults to
\code{NULL}.}

\item{return}{Logical, should the data be returned as a dataframe.
Defaults to \code{TRUE}.}
}
\value{
The data downloaded from the given url as a dataframe, exact format specified by data_trans_fn
}
\description{
Generic Get Data Function
}
\examples{

tb_burden <- get_data(url = "https://extranet.who.int/tme/generateCSV.asp?ds=estimates",
save_name = "TB_burden")

head(tb_burden)

}
\seealso{
get_tb_burden get_data_dict
}
