% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_dict.R
\name{get_data_dict}
\alias{get_data_dict}
\title{Get the WHO Data Dictionary for TB Data}
\usage{

  get_data_dict(url = "https://extranet.who.int/tme/generateCSV.asp?ds=dictionary",
  path = "data-raw", save = TRUE, save_name = "TB_data_dict",
  return = TRUE)
}
\arguments{
\item{url}{Character string, indicating  the url of the data
dictionary. Default is current url.}

\item{path}{Character string, indicating the folder to save the data into,
defaults to \code{data-raw}.}

\item{save}{Logical, should the dictionary be saved for reuse. Defaults to
\code{TRUE}.}

\item{save_name}{Character string, name to save dictionary under. Defaults to
\code{TB_data_dict}.}

\item{return}{Logical, should the data  dictionary be returned as a dataframe.
Defaults to \code{TRUE}.}
}
\value{
The WHO TB data dictionary as a tibble with 4 variables:
variable_name, dataset, code_list, definition.
}
\description{
Get the data dictionary for TB burden data from the WHO, see
\href{http://www.who.int/tb/country/data/download/en/}{here} for details.
}
\examples{

dict <- get_data_dict()

head(dict)

}
\seealso{
search_data_dict
}
