% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tb_burden.R
\name{get_tb_burden}
\alias{get_tb_burden}
\title{Get the WHO TB Burden Data}
\usage{

  get_tb_burden(url = "https://extranet.who.int/tme/generateCSV.asp?ds=estimates",
  download_data = FALSE, save = TRUE, burden_save_name = "TB_burden",
  path = "data-raw", return = TRUE, verbose = TRUE)
}
\arguments{
\item{url}{Character string, indicating  the url of the TB burden data.
Default is current url.}

\item{download_data}{Logical, defaults to \code{FALSE}. If not found locally should the data be
downloaded from the specified URL?}

\item{save}{Logical, should the data be saved for reuse. Defaults to
\code{FALSE}.}

\item{burden_save_name}{Character string, name to save the data under. Defaults to
\code{TB_burden}.}

\item{path}{Character string, indicating the folder to save the data into,
defaults to \code{data-raw}.}

\item{return}{Logical, should the data be returned as a dataframe.
Defaults to \code{TRUE}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should additional status and progress messages
be displayed.}
}
\value{
The WHO TB burden data as a tibble.
}
\description{
Get the TB burden data from the WHO, see
\href{http://www.who.int/tb/country/data/download/en/}{here} for details.
}
\examples{

tb_burden <- get_tb_burden(download_data = TRUE, save = TRUE)

head(tb_burden)

}
\seealso{
get_data search_data_dict
}
