% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTBinR.R
\docType{package}
\name{getTBinR-package}
\alias{getTBinR}
\alias{getTBinR-package}
\title{getTBinR: Get WHO Tuberculosis Data}
\description{
Facilitates easy import of analysis ready World Health Organisation Tuberculosis data and provides plotting functions for exploratory data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.samabbott.co.uk/getTBinR}
  \item \url{https://github.com/seabbs/getTBinR}
  \item Report bugs at \url{https://github.com/seabbs/getTBinR/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (0000-0001-8057-8037)

}
\keyword{internal}
