% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tb_burden.R
\name{map_tb_burden}
\alias{map_tb_burden}
\title{Map TB Burden}
\usage{
map_tb_burden(df = NULL, metric = "e_inc_100k", metric_label = NULL,
  countries = NULL, compare_to_region = FALSE, facet = NULL,
  year = 2016, trans = "identity", interactive = FALSE, ...)
}
\arguments{
\item{df}{Dataframe of TB burden data, as sourced by \code{\link[getTBinR]{get_tb_burden}}.
If not specified then will automatically source the WHO TB burden data, either locally if available
or directly from the WHO.}

\item{metric}{Character string specifying the metric to plot}

\item{metric_label}{Character string specifying the metric label to use.}

\item{countries}{A character string specifying the countries to plot.}

\item{compare_to_region}{Logical, defaults to \code{FALSE}. If \code{TRUE} all
countries that share a region with those listed in \code{countries} will be plotted.
Note that this will override settings for \code{facet}, unless it is set to "country".}

\item{facet}{Character string, the name of the variable to facet by.}

\item{year}{Numeric, indicating the year of data to map. Defaults to 2016.}

\item{trans}{A character string specifying the transform to use on the mapped metric. Defaults to no
transform ("identity"). Other options include log scaling ("log") and log base 10 scaling
("log10"). For a complete list of options see \code{ggplot2::continous_scale}.}

\item{interactive}{Logical, defaults to \code{FALSE}. If \code{TRUE} then an interactive plot is
returned.}

\item{...}{Additional parameters to pass to \code{\link[getTBinR]{get_tb_burden}}.}
}
\value{
A plot of TB Incidence Rates by Country
}
\description{
Map measures of TB burden by country by specifying a metric from the TB burden data.
Specify a country or vector of countries in order to map them (otherwise it will map all countries).
Various other options are available for tuning the plot further.
}
\examples{

map_tb_burden()

}
\seealso{
plot_tb_burden plot_tb_burden_overview get_tb_burden search_data_dict
}
