% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pointwise_tb_burden.R
\name{plot_pointwise_tb_burden}
\alias{plot_pointwise_tb_burden}
\title{Plot TB Burden by Country}
\usage{
plot_pointwise_tb_burden(df = NULL, metric = "e_inc_100k",
  metric_label = NULL, conf = c("_lo", "_hi"), countries = NULL,
  facet = NULL, scales = "fixed", interactive = FALSE, ...)
}
\arguments{
\item{df}{Dataframe of TB burden data, as sourced by \code{\link[getTBinR]{get_tb_burden}}.
If not specified then will automatically source the WHO TB burden data, either locally if available
or directly from the WHO.}

\item{metric}{Character string specifying the metric to plot}

\item{metric_label}{Character string specifying the metric label to use.
Defaults to \code{NULL}, which then uses the metric's definition as the label.}

\item{conf}{Character vector specifying the name variations to use to specify the upper
and lower confidence intervals. Defaults to c("_lo", "_hi"), if set to \code{NULL}
then no confidence intervals are shown.}

\item{countries}{A character string specifying the countries to plot.}

\item{facet}{Character string, the name of the variable to facet by.}

\item{scales}{Character string, see ?ggplot2::facet_wrap for details. Defaults to "fixed",
alternatives are "free_y", "free_x", or "free".}

\item{interactive}{Logical, defaults to \code{FALSE}. If \code{TRUE} then an interactive plot is
returned.}

\item{...}{Additional parameters to pass to \code{\link[getTBinR]{get_tb_burden}}.}
}
\value{
A plot of TB Incidence Rates by Country
}
\description{
Developent function
}
\examples{

tb_burden <- get_tb_burden()

tb_burden_eur <- filter(tb_burden, g_whoregion \%in\% "EUR")
sample_countries <- sample(unique(tb_burden$country), 10)
plot_pointwise_tb_burden(tb_burden, countries = sample_countries, scales = "free")

}
\seealso{
get_tb_burden search_data_dict
}
