% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tb_burden.R
\name{plot_tb_burden}
\alias{plot_tb_burden}
\title{Plot TB Burden by Country}
\usage{
plot_tb_burden(df, no_countries = NULL, facet = NULL, scales = "fixed")
}
\arguments{
\item{df}{Dataframe of TB burden data, as sourced by \code{\link[getTBinR]{get_tb_burden}}.}

\item{no_countries}{Numeric, the number of countries to plot}

\item{facet}{Character string, the name of the variable to facet by.}

\item{scales}{Character string, see ?ggplot2::facet_wrap for details. Defaults to "fixed",
alternatives are "free_y", "free_x", or "free".}
}
\value{
A plot of TB Incidence Rates by Country
}
\description{
Plot measures of TB burden by country. Currently only supports plotting of incidence rates.
In development so use with caution.
}
\examples{

tb_burden <- get_tb_burden()

plot_tb_burden(tb_burden, no_countries = 9, facet = "country")

}
