% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tb_burden_overview.R
\name{plot_tb_burden_overview}
\alias{plot_tb_burden_overview}
\title{Plot an overview of TB Burden for Multiple Countries}
\usage{
plot_tb_burden_overview(df = NULL, metric = "e_inc_100k",
  metric_label = NULL, countries = NULL, compare_to_region = FALSE,
  facet = NULL, scales = "free_y", interactive = FALSE, ...)
}
\arguments{
\item{df}{Dataframe of TB burden data, as sourced by \code{\link[getTBinR]{get_tb_burden}}.
If not specified then will automatically source the WHO TB burden data, either locally if available
or directly from the WHO.}

\item{metric}{Character string specifying the metric to plot}

\item{metric_label}{Character string specifying the metric label to use.}

\item{countries}{A character string specifying the countries to plot.}

\item{compare_to_region}{Logical, defaults to \code{FALSE}. If \code{TRUE} all
countries that share a region with those listed in \code{countries} will be plotted.
Note that this will override settings for \code{facet}, unless it is set to "country".}

\item{facet}{Character string, the name of the variable to facet by.}

\item{scales}{Character string, see ?ggplot2::facet_wrap for details. Defaults to "fixed",
alternatives are "free_y", "free_x", or "free".}

\item{interactive}{Logical, defaults to \code{FALSE}. If \code{TRUE} then an interactive plot is
returned.}

\item{...}{Additional parameters to pass to \code{\link[getTBinR]{get_tb_burden}}.}
}
\value{
A dot plot of any numeric metric by country.
}
\description{
This functions returns a dot plot for a given metric over a specified
list of countries. If \code{compare_to_region} is specified then a given country will
be compared to others in its region. This enables the user to rapidly understand trends in
Tuberculosis over time and the progress towards global elimination.
}
\examples{

plot_tb_burden_overview(countries = c("United Kingdom", "Botswana"), compare_to_region = TRUE)

}
\seealso{
get_tb_burden search_data_dict
}
