% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_df_plot.R
\name{prepare_df_plot}
\alias{prepare_df_plot}
\title{Generic Function to Prepare TB Burden Data for Plotting}
\usage{
prepare_df_plot(df = NULL, metric = "e_inc_100k", metric_label = NULL,
  countries = NULL, compare_to_region = FALSE, facet = NULL, ...)
}
\arguments{
\item{df}{Dataframe of TB burden data, as sourced by \code{\link[getTBinR]{get_tb_burden}}.
If not specified then will automatically source the WHO TB burden data, either locally if available
or directly from the WHO.}

\item{metric}{Character string specifying the metric to plot}

\item{metric_label}{Character string specifying the metric label to use.}

\item{countries}{A character string specifying the countries to plot.}

\item{compare_to_region}{Logical, defaults to \code{FALSE}. If \code{TRUE} all
countries that share a region with those listed in \code{countries} will be plotted.
Note that this will override settings for \code{facet}, unless it is set to "country".}

\item{facet}{Character string, the name of the variable to facet by.}

\item{...}{Additional parameters to pass to \code{\link[getTBinR]{get_tb_burden}}.}
}
\value{
A list containing 3 elements, the dataframe to plot, the facet to use and
the label to assign to the metric axis.
}
\description{
This function is used internally by \code{\link[getTBinR]{plot_tb_burden}}
and \code{\link[getTBinR]{plot_tb_burden_overview}} to prepare data for plotting.
}
\examples{

prepare_df_plot(countries = "Guinea")

}
\seealso{
plot_tb_burden plot_tb_burden_overview
}
