% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tb_dashboard.R
\name{run_tb_dashboard}
\alias{run_tb_dashboard}
\title{Run a TB Shiny Dashboard}
\usage{
run_tb_dashboard()
}
\value{
Starts a shiny Tuberculosis dashboard
}
\description{
This functions runs a TB dashboard that has been built using other
package functionality. The dashboard can be used to explore the global burden of TB
interactively. A hosted version of this dashboard is available \href{http://www.seabbs.co.uk/shiny/ExploreTB}{here}.
}
\examples{

## Only run the example if in an interative session
if (interactive()) {

## Run the TB dashboard
run_tb_dashboard()
}
}
