% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_who.R
\name{scale_colour_who}
\alias{scale_colour_who}
\alias{scale_color_who}
\title{Adds World Health Organization-inspired colours to plots}
\usage{
scale_colour_who(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{A character string. the colours of choice, current options are "main" and "light".}

\item{discrete}{Logical, defaults to \code{TRUE}. Should the palette be discrete or continuous.}

\item{reverse}{Logical, defaults to \code{FALSE}. Should the colour palette be reversed.}

\item{...}{Pass additional arguments to \code{ggplot2::discrete_scale} or \code{ggplot2::scale_colour_viridis_c}
depending on the \code{discrete} setting.}
}
\description{
Applies WHO inspired colours for \code{ggplot2} plots. Currently a continuous
palette has not been implemented so the \code{viridis} palettes will be used as a fallback in this situation.
}
\examples{

plot_tb_burden_summary(countries = "United Kingdom", compare_all_regions = FALSE,
                       compare_to_region = TRUE, conf = NULL, verbose = FALSE) +
theme_who() +
scale_colour_who(reverse = TRUE)
}
\seealso{
scale_colour_who scale_fill_who who_palettes
}
\author{
Maria Bekker-Nielsen Dunbar

Sam Abbott
}
