% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_who.R
\name{theme_who}
\alias{theme_who}
\title{World Health Organization-inspired ggplot2 theme}
\usage{
theme_who()
}
\description{
This theme is inspired by that used in the annual global Tuberculosis report.
See the report \href{https://www.who.int/tb/publications/global_report/en/}{here}.
}
\examples{

plot_tb_burden_summary(conf = NULL, verbose = FALSE) +
theme_who()
}
\seealso{
bbplot::bbc_style()
}
\author{
Maria Bekker-Nielsen Dunbar

Sam Abbott
}
