% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_palettes.R
\name{who_palettes}
\alias{who_palettes}
\title{World Health Organization-inspired palettes}
\usage{
who_palettes(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A character string. the colours of choice, current options are "main" and "light".}

\item{reverse}{Logical, defaults to \code{FALSE}. Should the palette be reversed.}

\item{...}{Pass additional arguments to \code{colorRampPalette}.}
}
\description{
World Health Organization-inspired palettes
}
\examples{

# Set up the main palette but reversed.
who_palettes(palette = "main", reverse = TRUE)
}
\seealso{
scale_colour_who scale_fill_who
}
\author{
Maria Bekker-Nielsen Dunbar

Sam Abbott
}
