% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_shapefile.R
\docType{data}
\name{who_shapefile}
\alias{who_shapefile}
\title{WHO shapefile}
\format{
A data frame with 15243 rows and 7 variables.
\describe{
\item{long}{Longitude}
\item{lat}{Latitude}
\item{order}{The shape order.}
\item{hole}{}
\item{piece}{}
\item{group}{The country group}
\item{id}{The country acronym}
}
}
\usage{
who_shapefile
}
\description{
The shapefile used in the WHO TB report, see \href{https://github.com/hazimtimimi/global_report/raw/master/Data/gparts.Rdata}{here}
for the original source. This shapefile is used in \code{\link[getTBinR]{map_tb_burden}}.
}
\keyword{datasets}
