% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTBinR.R
\docType{package}
\name{getTBinR-package}
\alias{getTBinR}
\alias{getTBinR-package}
\title{getTBinR: Get WHO Tuberculosis Data}
\description{
Quickly and easily import analysis ready Tuberculosis (TB) burden data, 
    from the World Health Organisation (WHO), into R. The aim of the package is to speed 
    up access to high quality TB burden data, using a simple R interface. Generic plotting
    functions are provided to allow for rapid graphical exploration of the WHO TB data. 
    A shiny dashboard is built in to showcase package functionality. An automated report
    allows country level TB burden to be summarised.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.samabbott.co.uk/getTBinR}
  \item \url{https://github.com/seabbs/getTBinR}
  \item Report bugs at \url{https://github.com/seabbs/getTBinR/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (0000-0001-8057-8037)

}
\keyword{internal}
