% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTBinR.R
\docType{package}
\name{getTBinR-package}
\alias{getTBinR}
\alias{getTBinR-package}
\title{getTBinR: Access and Summarise World Health Organisation Tuberculosis data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Quickly and easily import analysis ready
    Tuberculosis (TB) burden data, from the World Health Organisation
    (WHO), into R. The aim of getTBinR is to allow researchers, and other
    interested individuals, to quickly and easily gain access to a
    detailed TB data set and to start using it to derive key insights. It
    provides a consistent set of tools that can be used to rapidly
    evaluate hypotheses on a widely used data set before they are explored
    further using more complex methods or more detailed data. These tools
    include: generic plotting and mapping functions; a data dictionary
    search tool; an interactive shiny dashboard; and an automated, country
    level, TB report. For newer R users, this package reduces the barrier
    to entry by handling data import, munging, and visualisation. All
    plotting and mapping functions are built with ggplot2 so can be
    readily extended.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.samabbott.co.uk/getTBinR}
  \item \url{https://github.com/seabbs/getTBinR}
  \item Report bugs at \url{https://github.com/seabbs/getTBinR/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (0000-0001-8057-8037)

}
\keyword{internal}
