% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTBinR.R
\docType{package}
\name{getTBinR-package}
\alias{getTBinR}
\alias{getTBinR-package}
\title{getTBinR: Get WHO Tuberculosis Data in R}
\description{
The World Health Organisation compiles multiple datasets across infectious diseases. This package aims to facilitate easy import of analysis ready Tuberculosis data into R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.samabbott.co.uk/getTBinR}
  \item \url{https://github.com/seabbs/getTBinR}
  \item Report bugs at \url{https://github.com/seabbs/getTBinR/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk}

}
\keyword{internal}
