% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data_dict.R
\name{search_data_dict}
\alias{search_data_dict}
\title{Search the WHO TB Data Dictionary by Variable Name.}
\usage{
search_data_dict(var = NULL, verbose = TRUE, ...)
}
\arguments{
\item{var}{A character vector of variable names.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should search information be returned.}

\item{...}{Additional parameters to pass to \code{\link[getTBinR]{get_data_dict}}.}
}
\value{
A tibble containing the information in the data dictionary matching the variables
searched for.
}
\description{
Searches the WHO data dictionary for TB burden data. When run for the first time
it will download the data dictionary into the default path as specified by \code{\link[getTBinR]{get_data_dict}}.
To modify this behaviour pass arguements to \code{\link[getTBinR]{get_data_dict}}.
}
\examples{

search_data_dict(var = "country")

}
\seealso{
get_data_dict
}
